only use semver types on Web.
