+++ src/intellisense/notebookPythonPathService.node.ts
