+++ src/platform/common/utils.node.ts
@@ -4,6 +4,7 @@
 
 import * as path from '../../platform/vscode-path/path';
 import * as fsExtra from 'fs-extra';
+import { SemVer, parse } from 'semver';
 import { Uri } from 'vscode';
 import { fsPathToUri } from '../vscode-path/utils';
 import { IWorkspaceService } from './application/types';
@@ -69,4 +70,15 @@ export async function calculateWorkingDirectory(
         }
     }
     return workingDir;
+}
+
+// For the given string parse it out to a SemVer or return undefined
+export function parseSemVer(versionString: string): SemVer | undefined {
+    const versionMatch = /^\s*(\d+)\.(\d+)\.(.+)\s*$/.exec(versionString);
+    if (versionMatch && versionMatch.length > 2) {
+        const major = parseInt(versionMatch[1], 10);
+        const minor = parseInt(versionMatch[2], 10);
+        const build = parseInt(versionMatch[3], 10);
+        return parse(`${major}.${minor}.${build}`, true) ?? undefined;
+    }
 }
