+++ src/platform/common/utils.ts
@@ -4,7 +4,6 @@
 
 import type * as nbformat from '@jupyterlab/nbformat';
 import * as uriPath from '../../platform/vscode-path/resources';
-import { SemVer, parse } from 'semver';
 import { NotebookData, NotebookDocument, TextDocument, Uri, workspace } from 'vscode';
 import { sendTelemetryEvent } from '../../telemetry';
 import { getTelemetrySafeLanguage } from '../../telemetry/helpers';
@@ -153,17 +152,6 @@ export function generateNewNotebookUri(
         return Uri.joinPath(rootFolder ? Uri.file(rootFolder) : Uri.file(tmpDir), fileName).with({
             scheme: 'untitled'
         });
-    }
-}
-
-// For the given string parse it out to a SemVer or return undefined
-export function parseSemVer(versionString: string): SemVer | undefined {
-    const versionMatch = /^\s*(\d+)\.(\d+)\.(.+)\s*$/.exec(versionString);
-    if (versionMatch && versionMatch.length > 2) {
-        const major = parseInt(versionMatch[1], 10);
-        const minor = parseInt(versionMatch[2], 10);
-        const build = parseInt(versionMatch[3], 10);
-        return parse(`${major}.${minor}.${build}`, true) ?? undefined;
     }
 }
