+++ src/test/interpreters/index.node.ts
@@ -8,7 +8,7 @@ import { BufferDecoder } from '../../platform/common/process/decoder.node';
 import { PythonEnvInfo } from '../../platform/common/process/internal/scripts/index.node';
 import { ProcessService } from '../../platform/common/process/proc.node';
 import { PythonEnvironment } from '../../platform/pythonEnvironments/info';
-import { parsePythonVersion } from '../../platform/pythonEnvironments/info/pythonVersion';
+import { parsePythonVersion } from '../../platform/pythonEnvironments/info/pythonVersion.node';
 import { EXTENSION_ROOT_DIR_FOR_TESTS } from '../constants.node';
 import { isCondaEnvironment } from './condaLocator.node';
 import { getCondaEnvironment, getCondaFile, isCondaAvailable } from './condaService.node';
